feat: introduce macOS default settings
This file contains a bunch of default settings that we recommend to make
working with the OS more productive.

Please note that execution of this script is preventend for now with
`exit 0`. Remove this line to make this part of the initial setup.

Alternaively you can also run the script manually whenever you want.

Inspired by https://mths.be/macos